Eiffellib python library for Eiffel
Eiffellib is a library for subscribing to and publishing Eiffel events
to a message-broker.

This library solves the problem of publishing Eiffel events and
subscribing to events, removing the need of knowing how to connect
to a message-broker or how to utilize the protocol it supplies.

With Eiffellib you can start subscribing to and publish valid Eiffel
events quickly and to get a feel for the event protocol.

Change-Id: I57cdafae7da19ea1c059b7fa16c4bc4e3c02a2ec