Replace print() calls with the standard logging module (#17)
We should use the Python standard library's logging module to log
messages, not print() or traceback.print_exc(). While replacing
those calls we also adjust a few exception handlers to be able to
log better messages.

We also remove the otherwise unused EiffelSubscriber.verbose attribute
to let clients use the levels of the loggers to choose the logging
verbosity.