Specify open dependencies
Declaring fixed version numbers for requirements makes it hard to
impossible to use eiffellib in certain circumstances. Libraries should
always declare as open dependencies as possible to allow some
flexibility for the clients.

Should a client not want to allow for this flexibility they can always
use a constraints.txt (pip) or poetry.lock (poetry) to lock down the
version numbers for any transient dependencies. The other way around is
impossible, i.e allow a less restrictive version requirement and the
declared requirement from the package.

Increasing the jsonschema from an alpha version to the actual release
resolves #34. Requiring an alpha release when the actual release was
finalized over two years ago is sort of wicked.