Fix a problem where the RabbitMQPublisher could deadlock (#43)
There have been multiple problems with this code.
One problem is that it would send events from two different
threads. The main thread and the thread that resends nacked
events. This could cause the RabbitMQ server to disconnect
the client.
Secondly since the send_event method would block and wait
for the client to reconnect in case of disconnects, the
resend_nacked_events method could block the IOLoop which
would prevent the client from reconnecting, thus causing
a deadlock